% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{biplot_fa2}
\alias{biplot_fa2}
\title{biplot fa2 model}
\usage{
biplot_fa2(
  model,
  predictions,
  one = -1,
  second = -1,
  fscore = 2,
  fmult = 10,
  alpha = 0.3,
  alpha_site = 0.5,
  alpha_ind = 0.5,
  subtitle = NULL,
  gen = NULL,
  size_ind_biplot = 3
)
}
\arguments{
\item{model}{An asreml object with Factor analytic structure (fa2 or more)}

\item{predictions}{dataframe with predicted values by site}

\item{one}{c(1,-1)}

\item{second}{c(1,-1)}

\item{fscore}{score to visualize genotypes and avoid overlapping}

\item{fmult}{in order to see the sites and genotypes in the same plot we have to play with this value}

\item{alpha}{opacity segments}

\item{alpha_site}{opacity sites}

\item{alpha_ind}{opacity individuals}

\item{subtitle}{text}

\item{gen}{vector with genotypes to show (optional)}

\item{size_ind_biplot}{size_ind_biplot}
}
\value{
list with 3 ggplots
}
\description{
biplot fa2 model
}
\examples{
# library(tidyverse)
# library(asreml)
# library(agridat)
# data(besag.met)
# dat <- besag.met
#
# dat <- dat \%>\% arrange(county)
# model <- asreml(fixed = yield ~ 1 + county,
#                 random = ~ fa(county, 2):gen + county:rep + diag(county):rep:block,
#                 residual = ~ dsum(~ units | county),
#                 data = dat,
#                 na.action = list(x="include",y="include"))
#
# pp <- predict(model, classify = "county")$pvals

# sites
# p1 <- biplot_fa2(model,predictions = pp, fscore = 1, one = 1, second = 1, fmult = 5)$g1
# # sites scaled
# p2 <- biplot_fa2(model,predictions = pp, fscore = 1, one = 1, second = 1, fmult = 5)$g2
# # biplot
# p3 <- biplot_fa2(model,predictions = pp, fscore = 1, one = 1, second = 1, fmult = 5)$g3
}
