% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConection.R
\name{check_connectivity}
\alias{check_connectivity}
\title{Check connectivity between trials}
\usage{
check_connectivity(
  data = NULL,
  genotype = "line",
  trial = "Experiment",
  response = NULL,
  all = FALSE
)
}
\arguments{
\item{data}{A data.frame in a wide format.}

\item{genotype}{A character string indicating the column in data that
contains genotypes.}

\item{trial}{A character string indicating the column in data that contains
trials.}

\item{response}{A character string specifying the trait.}

\item{all}{Whether or not print all the table.}
}
\value{
Data frame with the genotype connectivity.
}
\description{
Check connectivity between trials
}
\examples{
\donttest{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
head(
  check_connectivity(
    data = dat,
    genotype = "gen",
    trial = "county",
    response = "yield",
    all = TRUE
  )
)
}
}
