% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDesign.R
\name{check_design_MET}
\alias{check_design_MET}
\title{Check Experimental Design}
\usage{
check_design_MET(
  data = NULL,
  genotype = NULL,
  trial = NULL,
  traits = NULL,
  rep = NULL,
  block = NULL,
  row = NULL,
  col = NULL
)
}
\arguments{
\item{data}{dataset}

\item{genotype}{string with genotype column}

\item{trial}{string with trial column}

\item{traits}{vector with traits}

\item{rep}{string with rep column}

\item{block}{string with block column}

\item{row}{string with row column}

\item{col}{string with col column}
}
\value{
list()
}
\description{
Check Experimental Design
}
\examples{
# In progress
}
