% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConection.R
\name{checkConection}
\alias{checkConection}
\title{checkconection}
\usage{
checkConection(
  data,
  genotype = "line",
  trial = "Experiment",
  response = "YDHA",
  all = FALSE
)
}
\arguments{
\item{data}{MET dataset}

\item{genotype}{string}

\item{trial}{string}

\item{response}{string}

\item{all}{wheater or not print all the table}
}
\value{
table with gen conection
}
\description{
checkconection
}
\examples{
# in process
}
