% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConection.R
\name{connectivity_matrix}
\alias{connectivity_matrix}
\title{connectivity_matrix}
\usage{
connectivity_matrix(
  data,
  genotype = "germplasmName",
  trial = "trial",
  response = NULL
)
}
\arguments{
\item{data}{MET dataset}

\item{genotype}{string}

\item{trial}{string}

\item{response}{string}
}
\value{
matrix
}
\description{
This function generates a (n_trial x n_trial) matrix  with the
 amount of genotypes shared between each pair of trial.
}
\examples{
# library(agridat)
# data(besag.met)
# dat <- besag.met
# conn <- connectivity_matrix(
#  data = dat,
#  genotype = "gen",
#  trial = "county",
#  response = "yield"
# )
# heatmap(conn)
# res <- factoextra::hcut(conn, k = 3, stand = FALSE)
# factoextra::fviz_dend(
#   x = res,
#   rect = FALSE,
#   cex = 0.5,
#   lwd = 0.5,
#   main = "Dendrogram",
#   horiz = TRUE
# )
}
