% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpATS.R
\name{CV_SpATS}
\alias{CV_SpATS}
\title{coefficient of variation SpATS}
\usage{
CV_SpATS(model)
}
\arguments{
\item{model}{SpATS object}
}
\value{
value
}
\description{
coefficient of variation SpATS
}
\examples{
# in progress
}
