% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractG.R
\name{extract_vcov}
\alias{extract_vcov}
\title{Extract Variance-Covariance from ASReml-R}
\usage{
extract_vcov(model = NULL, gen = "genotype", env = "trial", vc_model = "corv")
}
\arguments{
\item{model}{ASReml object}

\item{gen}{A character string indicating the column in data that contains
genotypes.}

\item{env}{A character string indicating the column in data that contains
environments or trials.}

\item{vc_model}{A character string indicating the variance-covariance fitted.
Can be 'diag', 'corv', 'corh', 'corgv', 'fa1', 'fa2', 'fa3', 'fa4', 'corgh',
'us' or 'rr2'.}
}
\value{
An object with a list of:
\item{VCOV}{A matrix of the estimated covariances between trials.}
\item{CORR}{A n_trial x n_trial matrix with the correlation between trials.}
\item{vc_model}{A character string indicating the variance-covariance fitted.}
}
\description{
Extract Variance-Covariance from ASReml-R
}
\examples{
\donttest{
library(agridat)
library(asreml)
library(agriutilities)

data(besag.met)
dat <- besag.met
dat <- dat \%>\% arrange(county)

model <- asreml(
 fixed = yield ~ 1 + county,
 random = ~ fa(county, 2):gen + county:rep + diag(county):rep:block,
 residual = ~ dsum(~ units | county),
 data = dat,
 na.action = list(x = "include", y = "include"),
 trace = 0
)
extract_vcov(model, gen = "gen" , env = "county", vc_model = "corv")
}
}
