% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractG.R
\name{extractG}
\alias{extractG}
\title{extractG}
\usage{
extractG(model, gen = "genotype", env = "trial", vc.model = "corv")
}
\arguments{
\item{model}{ASReml object}

\item{gen}{string with genotypes}

\item{env}{string with genotypes}

\item{vc.model}{variance covariance that you fitted}
}
\value{
list VCOV = VCOV , CORR = CORR, vc.model = vc.model
}
\description{
extractG
}
\examples{
# extractG(model, gen = "genotype" , env = "trial", vc.model = "corv")
}
