% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{fa2_summary}
\alias{fa2_summary}
\title{Factor Analytic Summary}
\usage{
fa2_summary(
  model = NULL,
  trial = "trial",
  genotype = "genotype",
  BLUEs_trial = NULL,
  mult_fa1 = -1,
  mult_fa2 = 1,
  filter_score = 1.5,
  k_biplot = 1,
  size_label_var = 2,
  alpha_label_var = 0.2,
  size_label_ind = 2,
  alpha_label_ind = 0.8,
  size_arrow = 0.2,
  alpha_arrow = 0.2,
  base_size = 12
)
}
\arguments{
\item{model}{factor analytic model (asreml object)}

\item{trial}{string}

\item{genotype}{string}

\item{BLUEs_trial}{data.frame with trial BLUEs}

\item{mult_fa1}{c(1,-1) (-1 by default)}

\item{mult_fa2}{c(1,-1) (1 by default)}

\item{filter_score}{value to filter genotypes by the distance from the origin}

\item{k_biplot}{factor to multiply the scores in the biplot}

\item{size_label_var}{double}

\item{alpha_label_var}{(0,1)}

\item{size_label_ind}{double}

\item{alpha_label_ind}{(0,1)}

\item{size_arrow}{double}

\item{alpha_arrow}{(0,1)}

\item{base_size}{double}
}
\value{
list with  loadings = L, loading_star,
Gvar, Cmat, summary_loadings, paf_site, var_tot, scores,
plots = list(loadings, biplot,  biplot_scaled,  loadings_c)
}
\description{
Factor Analytic Summary
}
\examples{
# library(tidyverse)
# library(asreml)
# library(agridat)
# data(besag.met)
# dat <- besag.met
#
# dat <- dat \%>\% arrange(county)
# model <- asreml(fixed = yield ~ 1 + county,
#                 random = ~ fa(county, 2):gen + county:rep + diag(county):rep:block,
#                 residual = ~ dsum(~ units | county),
#                 data = dat,
#                 na.action = list(x="include",y="include"))
#
# pp <- predict(model, classify = "county")$pvals
# fa2_summary(
#  model = model,
#  trial = "county",
#  genotype = "gen",
#  BLUEs_trial = pp,
#  mult_fa1 = -1,
#  mult_fa2 = -1,
#  filter_score = 1,
#  k_biplot = 10,
#  size_label_var = 3,
#  alpha_label_var = 0.5,
#  size_label_ind = 3,
#  alpha_label_ind = 0.8,
#  size_arrow = 0.2,
#  alpha_arrow = 0.1
# )
}
