% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{h.cullis}
\alias{h.cullis}
\title{Cullis heritability}
\usage{
h.cullis(model, gen)
}
\arguments{
\item{model}{lmer model}

\item{gen}{String genotype}
}
\value{
value
}
\description{
Cullis heritability
}
\examples{
# in progress
}
