% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{h_cullis}
\alias{h_cullis}
\title{Cullis heritability for lme4 model}
\usage{
h_cullis(model, genotype)
}
\arguments{
\item{model}{lmer model}

\item{genotype}{String genotype}
}
\value{
value
}
\description{
Cullis heritability for lme4 model
}
\examples{
# dat <- agridat::john.alpha
## fit model ---------------------------------------------------------------
## random genotype effect
# g.ran <- lme4::lmer(data  = dat, formula = yield ~ rep + (1|gen) + (1|rep:block))
# h_cullis(g.ran, "gen")
}
