% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{h_cullis_MME}
\alias{h_cullis_MME}
\title{Cullis heritability reconstructing mixed model equation}
\usage{
h_cullis_MME(model, genotype, returnMatrices = F)
}
\arguments{
\item{model}{lmer model}

\item{genotype}{String genotype}

\item{returnMatrices}{logical}
}
\value{
heritability or list with matrices
}
\description{
Cullis heritability reconstructing mixed model equation
}
\examples{
# dat <- agridat::john.alpha
## fit model ---------------------------------------------------------------
## random genotype effect
# g.ran <- lme4::lmer(data  = dat, formula = yield ~ rep + (1|gen) + (1|rep:block))
# h_cullis_MME(g.ran, "gen")
}
\author{
Paul Schmidt
}
