% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritability.R
\name{heritability_fa}
\alias{heritability_fa}
\title{Heritability for Factor Analytic Models in ASReml-R}
\usage{
heritability_fa(
  model_fa = NULL,
  genotype = "line",
  env = "loc",
  vc.model = c("fa2")
)
}
\arguments{
\item{model_fa}{Factor Analytic asreml model}

\item{genotype}{String}

\item{env}{String}

\item{vc.model}{variance covariance structure c("fa1", "fa2", "fa3", "fa4",
"us")}
}
\value{
data.frame
}
\description{
Heritability for Factor Analytic Models in ASReml-R
}
\examples{
# library(tidyverse)
# library(asreml)
# library(agridat)
# library(agriutilities)
# data(besag.met)
# dat <- besag.met
#
# dat <- dat \%>\% arrange(county)
# model <- asreml(
#   fixed = yield ~ 1 + county,
#   random = ~ fa(county, 2):gen + county:rep + diag(county):rep:block,
#   residual = ~ dsum(~ units | county),
#   data = dat,
#   na.action = list(x="include",y="include")
# )
# heritability_fa(
#  model_fa = model,
#  genotype = "gen",
#  env = "county",
#  vc.model = "fa2"
# )
}
