% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{lme4_res}
\alias{lme4_res}
\title{residuals lme4}
\usage{
lme4_res(model, returnN = F, k = 3)
}
\arguments{
\item{model}{lmer model}

\item{returnN}{logical value for returning the number of extreme observations}

\item{k}{number of standard desviations to define values as extremes}
}
\value{

}
\description{
residuals lme4
}
\examples{
# in progress
}
