% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{mult_lme4}
\alias{mult_lme4}
\title{Multiple lme4 models}
\usage{
mult_lme4(data, equation, var_sub)
}
\arguments{
\item{data}{MET data}

\item{equation}{formula see example below. You can use reformulate()}

\item{var_sub}{string variable for subset}
}
\value{

}
\description{
Multiple lme4 models
}
\examples{
# library(agridat)
# data(besag.met)
# dat <- besag.met
#
# # lme4 equation
# equation <- reformulate(c("rep", ran("gen") ), response = "yield")
#
# # fitting models
# objt <- mult_lme4(data = dat, equation = equation, var_sub = "county")
#
# # summary models
# mt_summ <- mult_summary(objt , genotype = "gen", y = "yield")
# mt_summ
#
# # outliers
# mult_res_lme4(models = objt)
#
# # removing outliers
# dataOut <- mult_res_lme4(models = objt, returnData = T)
#
# # refitting
# objt <- mult_lme4(data = dataOut, equation = equation, var_sub = "Experiment")
# mt_summ2 <- mult_summary(models = objt ,genotype =  "gen", y = "yield" )
}
