% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{mult_summary}
\alias{mult_summary}
\title{Summary for mult_lme4}
\usage{
mult_summary(models, genotype = "Name", y = "response", k = 3)
}
\arguments{
\item{models}{model from mult_lme4}

\item{genotype}{string genotype}

\item{y}{string response}

\item{k}{number of standard desviations to define values as extremes (default = 3)}
}
\value{
data.frame
}
\description{
Summary for mult_lme4
}
\examples{
# library(agridat)
# data(besag.met)
# dat <- besag.met
#
# # lme4 equation
# equation <- reformulate(c("rep", ran("gen") ), response = "yield")
#
# # fitting models
# objt <- mult_lme4(data = dat, equation = equation, var_sub = "county")
#
# # summary models
# mt_summ <- mult_summary(objt , genotype = "gen", y = "yield")
# mt_summ
#
# # outliers
# mult_res_lme4(models = objt)
#
# # removing outliers
# dataOut <- mult_res_lme4(models = objt, returnData = T)
#
# # refitting
# objt <- mult_lme4(data = dataOut, equation = equation, var_sub = "Experiment")
# mt_summ2 <- mult_summary(models = objt ,genotype =  "gen", y = "yield" )
}
