% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpATS.R
\name{plot_res_fitted}
\alias{plot_res_fitted}
\title{Interactive residuals vs fitted values SpATS}
\usage{
plot_res_fitted(data_out)
}
\arguments{
\item{data_out}{res_spats data.frame}
}
\value{
plotly
}
\description{
Interactive residuals vs fitted values SpATS
}
\examples{
# in progress
}
