% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpATS.R
\name{plot_res_index}
\alias{plot_res_index}
\title{Interactive residuals plot SpATS}
\usage{
plot_res_index(data_out)
}
\arguments{
\item{data_out}{res_spats data.frame}
}
\value{
plotly
}
\description{
Interactive residuals plot SpATS
}
\examples{
# in progress
}
