% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpATS.R
\name{plot_res_map}
\alias{plot_res_map}
\title{Interactive residuals plot in the field SpATS}
\usage{
plot_res_map(data_out)
}
\arguments{
\item{data_out}{res_spats data.frame}
}
\value{
plotly
}
\description{
Interactive residuals plot in the field SpATS
}
\examples{
# in progress
}
