% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{print.checkAgri}
\alias{print.checkAgri}
\title{Print an object of class \code{checkAgri}}
\usage{
\method{print}{checkAgri}(x, ...)
}
\arguments{
\item{x}{An object fitted with the function \code{check_design_met()}.}

\item{...}{Options used by the tibble package to format the output. See
`tibble::print()` for more details.}
}
\value{
an object inheriting from class \code{checkAgri}.
}
\description{
Prints information about \code{check_design_met()} function.
}
\examples{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
results <- check_design_met(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
print(results)
}
\author{
Johan Aparicio [aut]
}
