% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{res_lme4}
\alias{res_lme4}
\title{residuals lme4}
\usage{
res_lme4(model, returnN = F, k = 3)
}
\arguments{
\item{model}{lmer model}

\item{returnN}{logical value for returning the number of extreme observations}

\item{k}{number of standard desviations to define values as extremes}
}
\value{
data.frame of number of extreme observations depending on returnN
}
\description{
residuals lme4
}
\examples{
# in progress
}
