% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpATS.R
\name{res_spats}
\alias{res_spats}
\title{residuals SpATS}
\usage{
res_spats(model, k = 3)
}
\arguments{
\item{model}{SpATS object}

\item{k}{number of standard desviations to define values as extremes (default = 3)}
}
\value{
data.frame
}
\description{
residuals SpATS
}
\examples{
# in progress
}
