% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSingle.R
\name{single_model_analysis}
\alias{single_model_analysis}
\title{Single Trial Analysis}
\usage{
single_model_analysis(
  results = NULL,
  progress = TRUE,
  engine = "asreml",
  remove_outliers = TRUE
)
}
\arguments{
\item{results}{Object of class 'checkAgri' resulting of executing
check_design_MET function.}

\item{progress}{Should the progress of the modeling be printed.
If TRUE, for every trial a line is output indicating the traits fitted for
the particular trial.}

\item{engine}{A character string specifying the name of the mixed modeling
engine to use, either "lme4" or "asreml". For spatial designs, "SpaTS" is
always used, for other designs "asreml" as a default.}

\item{remove_outliers}{Should outliers be removed? TRUE by default.}
}
\value{
A list of data.frames.
}
\description{
Single Trial Analysis
}
\examples{
\donttest{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
results <- check_design_MET(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
out <- single_model_analysis(results, progress = FALSE)
out$resum_fitted_model
}
}
