% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMET.R
\name{stability}
\alias{stability}
\title{Stability Coefficients}
\usage{
stability(
  predictions = NULL,
  genotype = NULL,
  trial = NULL,
  response = NULL,
  best = "max"
)
}
\arguments{
\item{predictions}{A data.frame with one value per GxE combination.}

\item{genotype}{A character string indicating the column in data that
contains genotypes.}

\item{trial}{A character string indicating the column in data that contains
trials.}

\item{response}{A character vector specifying the response variable.}

\item{best}{A character string specifying how to define the best genotype
by numeric value ("min", "max"). "max" by default.}
}
\value{
A data.frame with several stability measures.
"superiority" (cultivar-superiority measure), "static" (Shukla's stability
variance) and "wricke" (Wricke's ecovalence)
}
\description{
Stability Coefficients
}
\examples{
# in progress
}
