% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{var_fa}
\alias{var_fa}
\title{Variance explained}
\usage{
var_fa(model)
}
\arguments{
\item{model}{An asreml object with Factor analytic structure (fa2)}
}
\value{
vector
}
\description{
Variance explained
}
\examples{
# in progress
}
