% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{varG.pvalue}
\alias{varG.pvalue}
\title{varG.pvalue}
\usage{
varG.pvalue(model, gen)
}
\arguments{
\item{model}{lmer model}

\item{gen}{string genotype}
}
\value{
pvalue significance of random component
}
\description{
varG.pvalue
}
\examples{
# dat <- agridat::john.alpha
## fit model ---------------------------------------------------------------
## random genotype effect
# g.ran <- lme4::lmer(data  = dat, formula = yield ~ rep + (1|gen) + (1|rep:block))
# varG.pvalue(g.ran, "gen")
}
