% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{VarG}
\alias{VarG}
\title{varG}
\usage{
VarG(model, comp)
}
\arguments{
\item{model}{lmer model}

\item{comp}{String random component to be extracted}
}
\value{

}
\description{
varG
}
\examples{
# Varg(model, "gen")
}
