% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{biplot_fa2}
\alias{biplot_fa2}
\title{biplot fa2 model}
\usage{
biplot_fa2(
  model,
  predictions,
  one = -1,
  second = -1,
  fscore = 2,
  fmult = 10,
  alpha = 0.3,
  alpha_site = 0.5,
  alpha_ind = 0.5,
  subtitle = NULL,
  gen = NULL,
  size_ind_biplot = 3
)
}
\arguments{
\item{model}{An asreml object with Factor analytic structure (fa2 or more)}

\item{predictions}{dataframe with predicted values by site}

\item{one}{c(1,-1)}

\item{second}{c(1,-1)}

\item{fscore}{score to visualize genotypes and avoid overlapping}

\item{fmult}{in order to see the sites and genotypes in the same plot we have to play with this value}

\item{alpha}{opacity segments}

\item{alpha_site}{opacity sites}

\item{alpha_ind}{opacity individuals}

\item{subtitle}{text}

\item{gen}{vector with genotypes to show (optional)}

\item{size_ind_biplot}{size_ind_biplot}
}
\value{
list with 3 ggplots
}
\description{
biplot fa2 model
}
\examples{
# in progress
}
