% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConection.R
\name{check_connectivity}
\alias{check_connectivity}
\title{check_connectivity}
\usage{
check_connectivity(
  data,
  genotype = "line",
  trial = "Experiment",
  response = NULL,
  all = FALSE
)
}
\arguments{
\item{data}{MET dataset}

\item{genotype}{string}

\item{trial}{string}

\item{response}{string}

\item{all}{wheater or not print all the table}
}
\value{
table with gen conection
}
\description{
check_connectivity
}
\examples{
# library(agridat)
# data(besag.met)
# dat <- besag.met
# check_connectivity(data = dat, genotype = "gen", trial = "county", response = "yield", all = T)
}
