% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDesign.R
\name{check_design_MET}
\alias{check_design_MET}
\title{Check Experimental Design}
\usage{
check_design_MET(
  data = NULL,
  genotype = NULL,
  trial = NULL,
  traits = NULL,
  rep = NULL,
  block = NULL,
  row = NULL,
  col = NULL
)
}
\arguments{
\item{data}{A data.frame in a wide format.}

\item{genotype}{A character string indicating the column in data that
contains genotypes.}

\item{trial}{A character string indicating the column in data that contains
trials.}

\item{traits}{A character vector specifying the traits for which the models
should be fitted.}

\item{rep}{A character string indicating the column in data that contains
replicates.}

\item{block}{A character string indicating the column in data that contains
sub blocks.}

\item{row}{A character string indicating the column in data that contains the
row coordinates}

\item{col}{A character string indicating the column in data that contains the
column coordinates.}
}
\value{
An object of class \code{checkAgri}, a list of data.frames.
}
\description{
Check Experimental Design
}
\examples{
\donttest{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
results <- check_design_MET(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
print(results)
}
}
