% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConection.R
\name{connectivity_matrix}
\alias{connectivity_matrix}
\title{Connectivity Matrix}
\usage{
connectivity_matrix(
  data = NULL,
  genotype = "germplasmName",
  trial = "trial",
  response = NULL
)
}
\arguments{
\item{data}{A data.frame in a wide format.}

\item{genotype}{A character string indicating the column in data that
contains genotypes.}

\item{trial}{A character string indicating the column in data that contains
trials.}

\item{response}{A character string specifying the trait.}
}
\value{
matrix
}
\description{
This function generates a (n_trial x n_trial) matrix  with the
amount of genotypes shared between each pair of trial.
}
\examples{
\donttest{
library(agridat)
data(besag.met)
dat <- besag.met
connectivity_matrix(
  data = dat,
  genotype = "gen",
  trial = "county",
  response = "yield"
)
}
}
