% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{covcor_heat}
\alias{covcor_heat}
\title{Correlation Covariance heatmap}
\usage{
covcor_heat(matrix, corr = TRUE, size = 4, digits = 3)
}
\arguments{
\item{matrix}{matrix}

\item{corr}{logical (TRUE default, correlation matrix)}

\item{size}{letter size}

\item{digits}{integer}
}
\value{
ggplot
}
\description{
Correlation Covariance heatmap
}
\examples{
# data(iris)
# M = cor(iris[,-5])
# covcor_heat(M, corr = T)
}
