% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractG.R
\name{extract_vcov}
\alias{extract_vcov}
\title{Extract Variance-Covariance from ASReml-R}
\usage{
extract_vcov(model = NULL, gen = "genotype", env = "trial", vc_model = "corv")
}
\arguments{
\item{model}{ASReml object}

\item{gen}{A character string indicating the column in data that contains
genotypes.}

\item{env}{A character string indicating the column in data that contains
environments or trials.}

\item{vc_model}{A character string indicating the variance-covariance fitted.
Can be 'diag', 'corv', 'corh', 'corgv', 'fa1', 'fa2', 'fa3', 'fa4', 'corgh',
'us' or 'rr2'.}
}
\value{
list VCOV = VCOV , CORR = CORR, vc_model = vc_model
}
\description{
Extract Variance-Covariance from ASReml-R
}
\examples{
# extract_vcov(model, gen = "genotype" , env = "trial", vc_model = "corv")
}
