% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{h_cullis}
\alias{h_cullis}
\title{Cullis heritability for lme4 models}
\usage{
h_cullis(model, genotype, re_MME = FALSE)
}
\arguments{
\item{model}{Object of class 'lmer'.}

\item{genotype}{A character string indicating the column in data that
contains genotypes.}

\item{re_MME}{A logical value to ask if we want to reconstruct the mixed models
equations to estimate the Cullis heritability. (FALSE by default)}
}
\value{
A numerical value of the Cullis heritability estimate. If re_MME is
TRUE, a list with matrices is returned.
}
\description{
Cullis heritability for lme4 models
}
\examples{
\donttest{
library(lme4)
library(agridat)
library(agriutilities)
# random genotype effect
dat <- agridat::john.alpha
g.ran <- lme4::lmer(
  data = dat,
  formula = yield ~ rep + (1 | gen) + (1 | rep:block)
)
h_cullis(model = g.ran, genotype = "gen")
}
}
\author{
Paul Schmidt
}
