% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritability.R
\name{heritability_fa}
\alias{heritability_fa}
\title{Heritability for Factor Analytic Models in ASReml-R}
\usage{
heritability_fa(
  model_fa = NULL,
  genotype = "line",
  env = "loc",
  vc_model = c("fa2"),
  diag = FALSE
)
}
\arguments{
\item{model_fa}{Factor Analytic ASReml model}

\item{genotype}{A character string indicating the column in data that
contains genotypes.}

\item{env}{A character string indicating the column in data that contains
environments or trials.}

\item{vc_model}{A character string indicating the variance-covariance
structure. Can be "fa1", "fa2", "fa3", "fa4" or "us".}

\item{diag}{\code{TRUE} or \code{FALSE} depending on the user if they want to
take the elements on the diagonal of the variance-covariance matrix or the
elements out of the diagonal to estimate the heritability. \code{FALSE} by
default.}
}
\value{
A list
}
\description{
Heritability for Factor Analytic Models in ASReml-R
}
\examples{
\donttest{
library(dplyr)
library(asreml)
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met

dat <- dat \%>\% arrange(county)
model <- asreml(
  fixed = yield ~ 1 + county,
  random = ~ fa(county, 2):gen + county:rep + diag(county):rep:block,
  residual = ~ dsum(~ units | county),
  data = dat,
  na.action = list(x = "include", y = "include"),
  trace = 0
)
heritability_fa(
   model_fa = model,
   genotype = "gen",
   env = "county",
   vc_model = "fa2"
 )
}
}
