% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMET.R
\name{met_analysis}
\alias{met_analysis}
\title{Multi-Environmental Trial Analysis}
\usage{
met_analysis(
  sma_output = NULL,
  h2_filter = 0.2,
  workspace = "1gb",
  trials_to_fit_fa = 4
)
}
\arguments{
\item{sma_output}{Object of class \code{smaAgri} resulting of executing
\code{single_trial_analysis()} function.}

\item{h2_filter}{Numeric value to filter trials with poor heritability.
0.2 by default.}

\item{workspace}{Sets the workspace for the core \code{REML} routines in the
form of a number optionally followed directly by a valid measurement unit.
"128mb" by default.}

\item{trials_to_fit_fa}{Number of trials necessary to fit a Factor Analytic
structure for the GxE interaction term. 4 by default.}
}
\value{
A list with a summary of the fitted models.
}
\description{
Multi-Environmental Trial Analysis
}
\examples{
\donttest{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
results <- check_design_MET(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
out <- single_trial_analysis(results, progress = FALSE)
met_results <- met_analysis(out)
print(met_results)
}
}
