% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersGG.R
\name{parameters_GG}
\alias{parameters_GG}
\title{Genetic Gain parameters}
\usage{
parameters_GG(model, trait = "trait")
}
\arguments{
\item{model}{linear regression model (lm object)}

\item{trait}{string with trait evaluated to be added in the data.frame}
}
\value{
data.frame with some parameters from the linear regression
(Slope, se_Slope, Intercept, r2, Pr(>F) and Genetic_Gain%)
}
\description{
Genetic Gain parameters
}
\examples{
# parameters_GG(model)
}
