% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.metAgri}
\alias{plot.metAgri}
\title{Plot an object of class \code{metAgri}}
\usage{
\method{plot}{metAgri}(x, type = c("corr", "cov"), ...)
}
\arguments{
\item{x}{An object inheriting from class \code{metAgri} resulting of
executing the function \code{met_analysis()}}

\item{type}{A character string specifiying the type of plot. "corr" or "cov".}

\item{...}{Further graphical parameters passed to \code{covcor_heat()}.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{metAgri}
}
\examples{
\dontrun{
library(agridat)
library(agriutilities)
data(besag.met)
dat <- besag.met
results <- check_design_met(
  data = dat,
  genotype = "gen",
  trial = "county",
  traits = c("yield"),
  rep = "rep",
  block = "block",
  col = "col",
  row = "row"
)
out <- single_trial_analysis(results, progress = FALSE)
met_results <- met_analysis(out, progress = FALSE)
print(met_results)
plot(met_results, type = "corr")
plot(met_results, type = "cov")
}
}
\author{
Johan Aparicio [aut]
}
